// src/main/resources/www2/src/api/api.ts
export interface HookRestApiView {
    /**
     * 
     * @type {number}
     * @memberof HookRestApiView
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof HookRestApiView
     */
    'name'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof HookRestApiView
     */
    'enabled'?: boolean;
    /**
     * 
     * @type {Set<HookTriggerRestApiView>}
     * @memberof HookRestApiView
     */
    'triggers'?: Set<HookTriggerRestApiView>;
    /**
     * 
     * @type {string}
     * @memberof HookRestApiView
     */
    'type': string;
}
/**
 * 
 * @export
 * @interface HookTriggerRestApiView
 */
export interface HookTriggerRestApiView {
    /**
     * 
     * @type {string}
     * @memberof HookTriggerRestApiView
     */
    'type'?: HookTriggerRestApiViewTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof HookTriggerRestApiView
     */
    'item'?: string;
}

export const HookTriggerRestApiViewTypeEnum = {
    PostTask: 'POST_TASK'
} as const;

export type HookTriggerRestApiViewTypeEnum = typeof HookTriggerRestApiViewTypeEnum[keyof typeof HookTriggerRestApiViewTypeEnum];
// ...
export interface WebHook extends HookRestApiView {
    /**
     * 
     * @type {string}
     * @memberof WebHook
     */
    'action'?: WebHookActionEnum;
    /**
     * 
     * @type {string}
     * @memberof WebHook
     */
    'url'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof WebHook
     */
    'sslValidation'?: boolean;
}

export const WebHookActionEnum = {
    Xml: 'POST_XML',
    Json: 'POST_JSON'
} as const;

export type WebHookActionEnum = typeof WebHookActionEnum[keyof typeof WebHookActionEnum];